requirements.txt
art==5.2
discord==1.7.3
openpyxl==3.0.7
psutil==5.8.0
python-telegram-bot==13.7
requests==2.25.1
schedule==1.1.0
selenium==3.141.0
simplejson==3.13.2
text2art==0.2.0
urllib3==1.26.6

